Create and commit offset map
- Offset map compression
- Run-length encoding algorithm
- Bitset encoding
- Dynamic compression algorithm selection
- Handle rebalancing
- Wait on thread Future, instead of a sleep loop
- Remove some sneaky throws
- Stabilise tests
- Faster, more reliable tests - 9s vs 2 for all core unit tests

+++ async-consumer-core/src/main/java/io/confluent/csid/utils/Range.java
@@ -13,19 +13,19 @@ import java.util.stream.IntStream;
  */
 public class Range implements Iterable<Integer> {
 
-    private int limit;
+    private long limit;
 
-    public Range(int limit) {
+    public Range(long limit) {
         this.limit = limit;
     }
 
-    public static Range range(int max) {
+    public static Range range(long max) {
         return new Range(max);
     }
 
     @Override
     public Iterator<Integer> iterator() {
-        final int max = limit;
+        final long max = limit;
         return new Iterator<Integer>() {
 
             private int current = 0;
@@ -58,7 +58,7 @@ public class Range implements Iterable<Integer> {
     }
 
     public IntStream toStream() {
-        return IntStream.range(0, limit);
+        return IntStream.range(0, (int)limit);
     }
 
     static IntStream rangeStream(int i) {
